image/test/mochitest/test_bug614392.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=614392
-->
<head>
  <title>Test for Bug 614392</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=614392">Mozilla Bug 614392</a>
<p id="display"></p>
<div id="content" style="display: none">
  <img src="damon.jpg">
</div>
<pre id="test">
<script type="application/javascript">
/** Test for Bug 614392**/

SimpleTest.waitForExplicitFinish();

window.onload = function() {
  var img = SpecialPowers.wrap(document.getElementsByTagName("img")[0]);
  var container = img
      .getRequest(SpecialPowers.Ci.nsIImageLoadingContent.CURRENT_REQUEST)
      .image;

  container.animationMode =
      SpecialPowers.Ci.imgIContainer.kDontAnimMode;

  is(container.animationMode,
     SpecialPowers.Ci.imgIContainer.kDontAnimMode,
     "yay, our animationMode tweak took effect (and we didn't crash!)");

  SimpleTest.finish();
}

</script>
</pre>
</body>
</html>
